+++ hw-3/simple-chat/src/index.js
@@ -2,3 +2,4 @@ import './index.css';
 import './screen1_style.css';
 import './screen2_style.css';
 import './screen1.js';
+import './screen2.js';
